move general components to src/app
